relocate registration page queries
